[refactor][볼링-step1] 모든 답변에 대해서, 다른 사람이 작성한 답변이 있는지 확인 코드 move(QnaService > Question)

+++ src/main/java/qna/service/QnAService.java
@@ -35,18 +35,13 @@ public class QnAService {
     @Transactional
     public void deleteQuestion(User loginUser, long questionId) throws CannotDeleteException {
         Question question = findQuestionById(questionId);
-
         question.verifyOwner(loginUser);
-
-        List<Answer> answers = question.getAnswers();
-        for (Answer answer : answers) {
-            answer.verifyOwner(loginUser);
-        }
+        question.verifyOwnerForAnswers(loginUser);
 
         List<DeleteHistory> deleteHistories = new ArrayList<>();
         question.setDeleted(true);
         deleteHistories.add(new DeleteHistory(ContentType.QUESTION, questionId, question.getWriter(), LocalDateTime.now()));
-        for (Answer answer : answers) {
+        for (Answer answer : question.getAnswers()) {
             answer.setDeleted(true);
             deleteHistories.add(new DeleteHistory(ContentType.ANSWER, answer.getId(), answer.getWriter(), LocalDateTime.now()));
         }
